Now able to have multiple *.dep.kts files in the project
The idea here is that for some reason, a repo may have very strict acls,
and different owners of the project may want to edit specific values. So
this would allow those owners to be able to split the definitions up
into multiple places giving them the ability to better 'acl' their
project.

This also removes the need to have the random file and load the classes
up from a detached configuration. I realized this won't work in most
cases as you won't have a repository set up early enough for this code
to take effect.